add component and scss
