+++ src/components/appbar/Appbar.module.scss
@@ -1,0 +1,18 @@
+@import "styles/variables";
+
+.header {
+  position: sticky;
+  top: 0;
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  background-color: $primary-color;
+  color: $secondary-color;
+  z-index: 101;
+  min-height: $header-height;
+  align-items: center;
+}
+
+.title {
+  font-size: 2em;
+}
