+++ src/components/appbar/Appbar.tsx
@@ -1,0 +1,9 @@
+import styles from './Appbar.module.scss'
+
+export default function Appbar() {
+  return (
+    <header className={styles.header}>
+      <div className={styles.title}>YUMEMI FE</div>
+    </header>
+  )
+}
