+++ src/components/footer/Footer.module.scss
@@ -1,0 +1,19 @@
+@import "styles/variables";
+
+.footer {
+  width: 100%;
+  height: 100px;
+  border-top: 1px solid #eaeaea;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-top:auto;
+  min-height: $footer-height;
+
+  a {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-grow: 1;
+  }
+}
