+++ src/pages/Home.module.scss
@@ -1,0 +1,48 @@
+@import "styles/variables";
+
+.home {
+  width: 100%;
+}
+
+.grid {
+  display: grid;
+  grid-template-columns: auto auto auto;
+  align-content: center;
+  justify-content: space-around;
+  &__item {
+
+  }
+}
+
+.chart {
+  margin-top: 2rem;
+  width: 100%;
+  height: 300px;
+}
+
+.checkbox {
+  -ms-transform: scale(1.5); /* IE */
+  -moz-transform: scale(1.5); /* FF */
+  -webkit-transform: scale(1.5); /* Safari and Chrome */
+  -o-transform: scale(1.5); /* Opera */
+  transform: scale(1.5);
+}
+
+@media (min-width: 800px) {
+  .chart {
+    width: 800px;
+    height: 300px;
+  }
+}
+
+@media (max-width: 150px) {
+  .chart {
+    width: 0px;
+  }
+}
+
+
+
+
+
+
