+++ src/pages/_app.tsx
@@ -1,8 +1,13 @@
 import '../../styles/globals.scss'
 import { appWithTranslation } from 'next-i18next'
+import Layout from '@components/layout/Layout'
 
 function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  )
 }
 
 export default appWithTranslation(MyApp)
