+++ styles/variables.scss
@@ -1,0 +1,5 @@
+$header-height: 80px;
+$footer-height: 60px;
+
+$primary-color: #000;
+$secondary-color: #fff;
